Refactored the code

+++ src/main/EventVariables.java
@@ -15,7 +15,7 @@ public class EventVariables {
 
     public static final int HORSE_MAX_STEP = 5;
 
-    public static final double INITIAL_WALLET = 500.0;
+    public static final int INITIAL_WALLET = 500;
 
     public static final String LOG_FILEPATH = "logs/AfternoonAtRaces";
 }
